+++ metadata.rb
@@ -4,6 +4,7 @@ maintainer_email 'me@cobus.io'
 license          'MIT'
 description      'Installs/Configures chef-wrapper-mesos'
 long_description 'Installs/Configures chef-wrapper-mesos'
-version          '0.1.2'
+version          '0.1.5'
 
+depends 'zookeeper', '~> 3.0.2'
 depends 'mesos', '~> 3.5.1'