+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.1.5
+
+Adding the slave recipe.
+
+# 0.1.3
+
+Adding the quorum value flag.
+
 # 0.1.0
 
 Initial release of chef-wrapper-mesos